+++ src/main/java/wooteco/subway/exception/SectionCreateException.java
@@ -1,4 +1,4 @@
-package wooteco.subway.utils.exception;
+package wooteco.subway.exception;
 
 public class SectionCreateException extends SubwayException {
