+++ src/main/java/wooteco/subway/exception/SectionDeleteException.java
@@ -1,4 +1,4 @@
-package wooteco.subway.utils.exception;
+package wooteco.subway.exception;
 
 public class SectionDeleteException extends SubwayException {
