+++ src/main/java/wooteco/subway/exception/SubwayException.java
@@ -1,4 +1,4 @@
-package wooteco.subway.utils.exception;
+package wooteco.subway.exception;
 
 public class SubwayException extends RuntimeException {
