+++ src/main/java/wooteco/subway/service/LineService.java
@@ -14,7 +14,7 @@ import wooteco.subway.dto.StationResponse;
 import wooteco.subway.repository.LineRepository;
 import wooteco.subway.repository.SectionRepository;
 import wooteco.subway.repository.StationRepository;
-import wooteco.subway.utils.exception.NameDuplicatedException;
+import wooteco.subway.exception.NameDuplicatedException;
 
 @Transactional
 @Service
