+++ src/test/java/wooteco/subway/service/LineServiceTest.java
@@ -25,7 +25,7 @@ import wooteco.subway.dto.LineResponse;
 import wooteco.subway.repository.LineRepository;
 import wooteco.subway.repository.SectionRepository;
 import wooteco.subway.repository.StationRepository;
-import wooteco.subway.utils.exception.NameDuplicatedException;
+import wooteco.subway.exception.NameDuplicatedException;
 
 @Transactional
 @SpringBootTest
